### 2023-06-14 #5560 @editorialbot
[PRE REVIEW]: Zoomerjoin: Superlatively Fast Fuzzy-Joins
**Submitting author:** @beniaminogreen (<a href="http://orcid.org/0009-0006-4501-597X">Beniamino Green</a>)
**Repository:** https://github.com/beniaminogreen/zoomerjoin
**Branch with paper.md** (empty if default branch): joss
**Version:** 0.1.0
