### 2023-06-14 041b499
fixed typo, added citation for rextendr
