+++ paper.bib
@@ -1,3 +1,10 @@
+@Manual{rextendr,
+    title = {rextendr: Call Rust Code from R using the 'extendr' Crate},
+    author = {Claus O. Wilke and Andy Thomason and Mossa M. Reimert and Ilia Kosenkov and Hiroaki Yutani and Malcolm Barrett},
+    year = {2023},
+    note = {R package version 0.3.0},
+    url = {https://CRAN.R-project.org/package=rextendr},
+}
 @misc{rayon,
 	title={rayon: Simple work-stealing parallelism for Rust},
 	author={Niko Matsakis and Josh Stone},
