### 2023-06-14 #5560 @beniaminogreen
Thanks! Looking at the list of reviewers I would suggest lmullen, as he has written a package (textreuse) that also uses Locality-Sensitive Hashing which inspired me to make this package. Beyond this no real opinions, although I think pmyteh, cjbarrie, and justinchuntngho might be able to supply good reviews.
