+++ pkgdown.yml
@@ -4,7 +4,7 @@ pkgdown_sha: ~
 articles:
   benchmarks: benchmarks.html
   guided_tour: guided_tour.html
-last_built: 2023-06-02T17:00Z
+last_built: 2023-06-18T16:55Z
 urls:
   reference: https://beniaminogreen.github.io/zoomerjoin/reference
   article: https://beniaminogreen.github.io/zoomerjoin/articles
