### 2023-06-18 ddf7f5f
regex_join prototype
