+++ R/extendr-wrappers.R
@@ -1,4 +1,7 @@
 # Generated by extendr: Do not edit by hand
+
+# nolint start
+
 #
 # This file was created with the following call:
 #   .Call("wrap__make_zoomerjoin_wrappers", use_symbols = TRUE, package_name = "zoomerjoin")
@@ -18,3 +21,7 @@ rust_em_link <- function(x_robj, probs, tol, max_iter) .Call(wrap__rust_em_link,
 
 rust_p_norm_join <- function(a_mat, b_mat, radius, band_width, n_bands, r) .Call(wrap__rust_p_norm_join, a_mat, b_mat, radius, band_width, n_bands, r)
 
+rust_regex_join <- function(left_string_r, `_right_string_r`, r_regex) .Call(wrap__rust_regex_join, left_string_r, `_right_string_r`, r_regex)
+
+
+# nolint end
