+++ src/rust/src/lib.rs
@@ -4,6 +4,11 @@ use extendr_api::prelude::*;
 use dashmap::{DashMap, DashSet};
 
 use rayon::prelude::*;
+use regex::Regex;
+
+use rustc_hash::FxHasher;
+use std::collections::HashMap;
+use std::hash::{Hash, Hasher};
 
 pub mod shingleset;
 use crate::shingleset::ShingleSet;
@@ -85,6 +90,67 @@ fn rust_jaccard_join(
 }
 
 #[extendr]
+fn rust_regex_join(left_string_r: Robj, _right_string_r: Robj, r_regex: &str) -> Robj {
+    let left_string_vec = left_string_r.as_str_vector().unwrap();
+    // let right_string_vec = right_string_r.as_str_vector().unwrap();
+
+    let re = Regex::new(r_regex).expect("Can't compile regular expression!");
+
+    let mut small_dict: HashMap<u64, Vec<u64>> = HashMap::new();
+
+    for (i, row) in left_string_vec.iter().enumerate() {
+        // let matches = re.find_iter(row).map(|mat| mat.as_str()).collect::<Vec<&str>>();
+        let mut hasher = FxHasher::default();
+
+        for re_match in re.find_iter(row).map(|mat| mat.as_str()) {
+            re_match.hash(&mut hasher)
+        }
+
+        let key = hasher.finish();
+
+        if key != 0 {
+            match small_dict.get_mut(&key) {
+                Some(matches) => matches.push(i as u64 +1),
+                None => {small_dict.insert(key, vec![i as u64 +1]);},
+            }
+        }
+    }
+
+    let mut output_vec: Vec<(u64, u64)> = Vec::new();
+
+    for (j, row) in left_string_vec.iter().enumerate() {
+        // let matches = re.find_iter(row).map(|mat| mat.as_str()).collect::<Vec<&str>>();
+        let mut hasher = FxHasher::default();
+
+        for re_match in re.find_iter(row).map(|mat| mat.as_str()) {
+            re_match.hash(&mut hasher)
+        }
+
+        let key = hasher.finish();
+
+        if key != 0 {
+            // relies on property of fx hasher (will be zero if nothing has been hashed)
+            match small_dict.get(&key) {
+                Some(matches) => {
+                    for i in matches {
+                        output_vec.push((*i, j as u64 + 1))
+                    }
+                }
+                None => {}
+            };
+        }
+    }
+
+    let mut out_arr: Array2<u64> = Array2::zeros((output_vec.len(), 2));
+    for (i, pair) in output_vec.iter().enumerate() {
+        out_arr[[i, 0]] = pair.1 as u64;
+        out_arr[[i, 1]] = pair.0 as u64;
+    }
+
+    Robj::try_from(&out_arr).into()
+}
+
+#[extendr]
 fn rust_salted_jaccard_join(
     left_string_r: Robj,
     right_string_r: Robj,
@@ -100,7 +166,6 @@ fn rust_salted_jaccard_join(
 
     let right_salt_vec = right_salt_r.as_str_vector().unwrap();
     let left_salt_vec = left_salt_r.as_str_vector().unwrap();
-
 
     let joiner = MinHashJoiner::new_with_salt(
         left_string_vec,
@@ -122,7 +187,14 @@ fn rust_salted_jaccard_join(
 }
 
 #[extendr]
-fn rust_p_norm_join(a_mat: Robj, b_mat: Robj, radius: f64, band_width : u64, n_bands: u64, r : f64) -> Robj {
+fn rust_p_norm_join(
+    a_mat: Robj,
+    b_mat: Robj,
+    radius: f64,
+    band_width: u64,
[…]
