+++ man/regex_right_join.Rd
@@ -1,0 +1,20 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/regex_logical_joins.R
+\name{regex_right_join}
+\alias{regex_right_join}
+\title{regex right-join using minihashing}
+\usage{
+regex_right_join(a, b, regex, by = NULL)
+}
+\arguments{
+\item{a}{the first dataframe you wish to join.}
+
+\item{b}{the second dataframe you wish to join.}
+
+\item{by}{a named vector indicating which columns to join on. Format should
+be the same as dplyr: \code{by = c("column_name_in_df_a" =
+"column_name_in_df_b")}}
+}
+\description{
+regex right-join using minihashing
+}
