+++ src/rust/Cargo.toml
@@ -18,4 +18,5 @@ nohash-hasher = "0.2.0"
 ordered-float = "3.4.0"
 rand = "0.8.5"
 rayon = "1.6.1"
+regex = "1.8.4"
 rustc-hash = "1.1.0"
