### 2023-06-19 aebc806
removed pesky print statement
