### 2023-08-02 #5693 @cjbarrie
This is a very good package. Easy to use and well written. There are just a few small changes I would recommend.

1. You do not provide the code you used to arrive at the benchmarks. It would be helpful to have this to test the claims of the article.
2. You do not attach data in the package that you use (corpus_1 and corpus_2) to demonstrate the package functionality. I think some version of this could be attached within R package file size constraints. If not, just provide a link as Wickham has done for e.g., `babynames` package in the past.
3. It would be helpful to have vignette documentation for the other package functions. The documentation remains quite minimal at the moment.
4. There are no Community guidelines I can see on the Github repo.

Additionally, some small typos in the article:

- In Summary section: "typically runs in runs in linear"
- In Statement of Need section: " is probabilistic a probabilistic algorithm"
- In Other Functionalities section: "implementation the probabilistic record-linkage"
- In Other Functionalities section: "to pairs to units"
