### 2023-08-02 #5693 @beniaminogreen
Hi,

Thanks for your detailed feedback and your kind words about the package.

Writing in response to the four points you raise:

1) Thanks for flagging the lack of the benchmarking code. I originally included this in the package, but it was accidentally deleted at some stage. I will include it at the top of the benchmarking vignette so that it's easily available to you and other users.
Unfortunately, the benchmarking takes something like 6 hours to run on a 8-core server. Most of this time is spent running the fuzzyjoin code, so it might be practical to only run the zoomerjoin code to verify its performance. Alternatively, I can reduce the size of the datasets or the number of trials to make sure it is replicable in a reasonable amount of time.

2) I have the corpus data in the vignettes file, but I am not sure if this is the right place. I will read through the R package book and make sure that I put the files in an appropriate place, and that the files are visible to users. I also very much like the code snippet that you provided, so I will include it or something similar.

3) I agree that the package should have better documentation for the Euclidean distance joins, and the fuzzy string grouping, which I will add in the next few days.

4) I have included a contributing guide / community code of conduct in the [contributing.md](https://github.com/beniaminogreen/zoomerjoin/blob/main/CONTRIBUTING.md) file in the root of the repo. Not sure if this is the right place, or the right format for the guidelines. I'll change it in the next commit if this does not meet the requirements.

I'll also make sure to fix the typos you flag in the paper.

Best and many thanks,
Ben
