+++ data/data.R
@@ -1,0 +1,15 @@
+#' Donors from DIME Database
+#'
+#' A sample of donors from the DIME database of political donors. Used as
+#' sample data. Also used in the "Adaptive Fuzzy String Matching: How to Merge
+#' Data Sets with Only One (Messy) Identifying Field" by Kaufman and Klevs."
+#'
+#'
+#' @format ## `bonica_data`
+#' A data frame with 1,332,460 rows and 2 columns:
+#' \describe{
+#'   \item{id}{Numeric ID / Row Number}
+#'   \item{x}{Donor Name}
+#'   ...
+#' }#' @source <https://www.who.int/teams/global-tuberculosis-programme/data>
+"bonica_data"
