+++ README.md
@@ -169,26 +169,26 @@ join_out <- jaccard_inner_join(corpus_1, corpus_2, n_gram_width=6, n_bands=20, b
 print(Sys.time() - start_time)
 ```
 
-    ## Time difference of 10.00245 secs
+    ## Time difference of 10.06169 secs
 
 ``` r
 print(join_out)
 ```
 
-    ## # A tibble: 188,352 × 4
-    ##         a field.x                                               b field.y       
-    ##     <dbl> <chr>                                             <dbl> <chr>         
-    ##  1    196 committee for responsible government             868477 tmha committe…
-    ##  2  70987 the law offices of lucio g. valdez               918658 law offices o…
-    ##  3  57366 grande communications networks inc. pac account  963865 grande commun…
-    ##  4 257012 esg international inc                           1002173 acg internati…
-    ##  5  68222 ball and weed                                    886213 ball and weed…
-    ##  6  95976 health care center                              1273955 lake health c…
-    ##  7 128542 mid valley democratic club                       899639 tri valley de…
-    ##  8 121580 steelworkers local 760                          1198389 steelworkers …
-    ##  9 268575 mdsa strategic communications                   1196490 strategic com…
-    ## 10 441473 sw construction inc                             1260268 mat construct…
-    ## # ℹ 188,342 more rows
+    ## # A tibble: 186,691 × 4
+    ##         a field.x                                                      b field.y
+    ##     <dbl> <chr>                                                    <dbl> <chr>  
+    ##  1 406569 west jersey optometric society                          1.20e6 south …
+    ##  2 302206 solutions inc                                           1.21e6 r-solu…
+    ##  3 254858 g f thomas investments lp                               1.19e6 thomas…
+    ##  4 362272 olive construction management inc                       1.16e6 cb con…
+    ##  5 237140 fernandez liquors inc                                   9.68e5 fernan…
+    ##  6  31191 anson county democratic party                           8.64e5 slo co…
+    ##  7 366593 lcg investment co lp                                    1.24e6 p2 inv…
+    ##  8  57768 united food and commercial workers international union… 8.89e5 united…
+    ##  9 408887 wdc enterprises inc                                     1.23e6 plc en…
+    ## 10 483822 r energy management co                                  1.00e6 acp en…
+    ## # ℹ 186,681 more rows
 
 ZoomerJoin finds and joins on the matching rows in just a few seconds.
 
