+++ vignettes/guided_tour.Rmd
@@ -16,10 +16,11 @@ library(fuzzyjoin)
 library(zoomerjoin)
 
 n <- 500000
-corpus_1 <- read_csv("bonica.csv")  %>%
+# bonica_data packaged with zoomerjoin
+corpus_1 <- bonica_data  %>%
     head(n)
 names(corpus_1) <- c("a", "field")
-corpus_2 <- read_csv("bonica.csv")  %>%
+corpus_2 <- bonica_data  %>%
     tail(n)
 names(corpus_2) <- c("b", "field")
 
