+++ README.Rmd
@@ -12,10 +12,10 @@ library(fuzzyjoin)
 devtools::load_all()
 
 n <- 500000
-corpus_1 <- read_csv("vignettes/bonica.csv")  %>%
+corpus_1 <- dime_data %>%
     head(n)
 names(corpus_1) <- c("a", "field")
-corpus_2 <- read_csv("vignettes/bonica.csv")  %>%
+corpus_2 <- dime_data %>%
     tail(n)
 names(corpus_2) <- c("b", "field")
 ```
