### 2023-08-04 c67fc88
added documentation of string_group to guided tour vignette
