### 2023-08-05 b3fa385
upgraded guided tour and matching on vectors
