+++ pkgdown.yml
@@ -4,7 +4,8 @@ pkgdown_sha: ~
 articles:
   benchmarks: benchmarks.html
   guided_tour: guided_tour.html
-last_built: 2023-08-05T03:38Z
+  matching_vectors: matching_vectors.html
+last_built: 2023-08-06T17:12Z
 urls:
   reference: https://beniaminogreen.github.io/zoomerjoin/reference
   article: https://beniaminogreen.github.io/zoomerjoin/articles
