### 2023-08-06 #5693 @beniaminogreen
Writing to share some of the updates to the package over the past few days.  In addition to fixing typos and fixing a bug in the euclidean joining code, I have:

1) Updated the [guided tour vingette](http://beniamino.org/zoomerjoin/articles/guided_tour.html) to show off the `jaccard_string_group` and `jaccard_curve` functions, as well as adding a [very short vingette](http://beniamino.org/zoomerjoin/articles/matching_vectors.html) on joining using the euclidean distance.

2) added the DIME / Bonica data so it is packaged with the R repository, which makes the functionality of the package easier to experiment. This change is reflected in the README and vignettes.

3) The benchmarking vingette now contains the code that is used to generate the benchmarks (although it isn't run every time). As mentioned before, this code takes several hours to run due to the large size of the datasets used.

Because of this, I have also uploaded [this gist](https://gist.github.com/beniaminogreen/ce284a7fcb59d7d759f7f8cdc8b161f6) which runs the same benchmarking code over a smaller set of inputs to show the difference in scaling between the two methods for smaller datasets. It should run in ~20 mins on a modern laptop.
