+++ pkgdown.yml
@@ -5,7 +5,7 @@ articles:
   benchmarks: benchmarks.html
   guided_tour: guided_tour.html
   matching_vectors: matching_vectors.html
-last_built: 2023-08-06T17:12Z
+last_built: 2023-08-08T13:27Z
 urls:
   reference: https://beniaminogreen.github.io/zoomerjoin/reference
   article: https://beniaminogreen.github.io/zoomerjoin/articles
