### 2023-08-08 beniaminogreen/zoomerjoin#70 @wincowgerDEV
Error in installation instructions
**Describe the bug**
I am running Windows 11 with rust 1.71.1 and wsl, when I try to install the package this is my console with an error message:

```r
> library(zoomerjoin)
Error in library(zoomerjoin) : there is no package called ‘zoomerjoin’
> ## Install with devtools
> # install.packages("devtools")
> devtools::install_github("beniaminogreen/zoomerjoin")
Downloading GitHub repo beniaminogreen/zoomerjoin@HEAD
These packages have more recent versions available.
It is recommended to update all of them.
Which would you like to update?

1: All
2: CRAN packages only
3: None
4: vctrs (0.6.2 -> 0.6.3) [CRAN]

Enter one or more numbers, or an empty line to skip updates: 3
── R CMD build ───────────────────────────────────────────────────────────────────────────────────────────────────────────
✔  checking for file 'C:\Users\winco\AppData\Local\Temp\RtmpCwBEaE\remotes522c7688750f\beniaminogreen-zoomerjoin-f359f28/DESCRIPTION' ...
─  preparing 'zoomerjoin':
─  cleaning src
─  building 'zoomerjoin_0.0.0.9000.tar.gz'

Installing package into ‘C:/Users/winco/AppData/Local/R/win-library/4.3’
(as ‘lib’ is unspecified)
* installing *source* package 'zoomerjoin' ...
** using staged installation
** libs
using C compiler: 'gcc.exe (GCC) 12.2.0'
rm -Rf zoomerjoin.dll ./rust/target/x86_64-pc-windows-gnu/release/libzoomerjoin.a entrypoint.o
gcc  -I"C:/PROGRA~1/R/R-43~1.0/include" -DNDEBUG     -I"C:/RBuildTools/4.3/x86_64-w64-mingw32.static.posix/include"     -O2 -Wall  -mfpmath=sse -msse2 -mstackrealign  -c entrypoint.c -o entrypoint.o
mkdir -p ./rust/target/libgcc_mock
cd ./rust/target/libgcc_mock && \
	touch gcc_mock.c && \
	gcc -c gcc_mock.c -o gcc_mock.o && \
	ar -r libgcc_eh.a gcc_mock.o && \
	cp libgcc_eh.a libgcc_s.a
C:\RBuildTools\4.3\x86_64-w64-mingw32.static.posix\bin\ar.exe: creating libgcc_eh.a
# CARGO_LINKER is provided in Makevars.ucrt for R >= 4.2
export CARGO_TARGET_X86_64_PC_WINDOWS_GNU_LINKER="x86_64-w64-mingw32.static.posix-gcc.exe" && \
	export LIBRARY_PATH="${LIBRARY_PATH};/c/Users/winco/AppData/Local/Temp/Rtmpctc8EF/R.INSTALL1c04338dae1/zoomerjoin/src/./rust/target/libgcc_mock" && \
	cargo +stable-gnu build --target=x86_64-pc-windows-gnu --lib --release --manifest-path=./rust/Cargo.toml --target-dir ./rust/target
error: toolchain 'stable-x86_64-pc-windows-gnu' is not installed
make: *** [Makevars.win:19: rust/target/x86_64-pc-windows-gnu/release/libzoomerjoin.a] Error 1
ERROR: compilation failed for package 'zoomerjoin'
* removing 'C:/Users/winco/AppData/Local/R/win-library/4.3/zoomerjoin'
Warning message:
In i.p(...) :
  installation of package ‘C:/Users/winco/AppData/Local/Temp/RtmpCwBEaE/file522c5d5356f/zoomerjoin_0.0.0.9000.tar.gz’ had non-zero exit status
```

I found this guidance https://github.com/rust-lang/rustup/issues/1793 and tried to follow along but the package is still not installing.

This is the result of rustup show for me:

```
Default host: x86_64-unknown-linux-gnu
rustup home:  /home/wincowger/.rustup

stable-x86_64-unknown-linux-gnu (default)
rustc 1.71.1 (eb26296b5 2023-08-03)
```

**To Reproduce**
Steps to reproduce the behavior:
Follow user guidelines on Windows 11 machine with WSL.

**Expected behavior**
A clear and concise description of what you expected to happen.
Installation to proceed.

**Screenshots**
If applicable, add screenshots to help explain your problem.

 - OS: Windows 11

**Additional context**
Add any other context about the problem here.
