### 2023-08-08 beniaminogreen/zoomerjoin#70 @beniaminogreen
Thanks for filing this! I do not have experience installing the package on a Windows computer with WSL, but I have had to complete extra steps when installing on a Windows server I have access to.  Generally, I have to make sure that the correct rust toolchains are installed, which doesn't seem to be the case here (although I don't have much experience using Windows).

I don't mean to suggest that you run something you've already tried, but can you try running the following commands?

```
rustup toolchain remove stable-x86_64-pc-windows-msvc
rustup install stable-x86_64-pc-windows-msvc
rustup toolchain remove stable-x86_64-pc-windows-gnu
rustup install stable-x86_64-pc-windows-gnu
```

Here's the output of my rustup-show on the windows server I have access to:

```
rustup show
Default host: x86_64-pc-windows-msvc
rustup home:  C:\Users\bfg9\.rustup

installed toolchains
--------------------

stable-x86_64-pc-windows-gnu
stable-x86_64-pc-windows-msvc (default)
nightly-x86_64-pc-windows-msvc

installed targets for active toolchain
--------------------------------------

i686-pc-windows-gnu
x86_64-pc-windows-gnu
x86_64-pc-windows-msvc

active toolchain
----------------

stable-x86_64-pc-windows-msvc (default)
rustc 1.70.0 (90c541806 2023-05-31)
```

I can see Windows installation as being a stumbling block for a lot of users, so I will make sure to update the instructions in the README after I discover a reproducible fix.

Best, Ben
