### 2023-08-08 #5693 @wincowgerDEV
@beniaminogreen recommendation on line 77 of the manuscript, I don't think you want to claim that your package never creates false positives. True positive matches in character datasets are often subjective and may be complex (beyond the capabilities of jacard similarity), it may confuse readers when they do not get the expected output.
