### 2023-08-08 #5693 @wincowgerDEV
@beniaminogreen, I recommend adding some code examples in text to demonstrate your points, it will help readers understand which function to use to achieve which outcomes you are talking about.
