### 2023-08-08 #5693 @wincowgerDEV
@beniaminogreen, it isn't clear where the code for producing the euclidean synthetic dataset is described in the manuscript.
