### 2023-08-08 beniaminogreen/zoomerjoin#70 @wincowgerDEV
Thanks for this. Definitely getting closer, may need to change the active toolchain.:

```
wincowger@Win:/mnt/c/WINDOWS/system32$ rustup install stable-x86_64-pc-windows-msvc
error: DEPRECATED: future versions of rustup will require --force-non-host to install a non-host toolchain as the default.
warning: toolchain 'stable-x86_64-pc-windows-msvc' may not be able to run on this system.
warning: If you meant to build software to target that platform, perhaps try `rustup target add x86_64-pc-windows-msvc` instead?
info: syncing channel updates for 'stable-x86_64-pc-windows-msvc'
info: latest update on 2023-08-03, rust version 1.71.1 (eb26296b5 2023-08-03)
info: downloading component 'cargo'
info: downloading component 'clippy'
info: downloading component 'rust-docs'
info: downloading component 'rust-std'
info: downloading component 'rustc'
info: downloading component 'rustfmt'
info: installing component 'cargo'
info: installing component 'clippy'
info: installing component 'rust-docs'
 13.6 MiB /  13.6 MiB (100 %)  10.5 MiB/s in  1s ETA:  0s
info: installing component 'rust-std'
 23.6 MiB /  23.6 MiB (100 %)  18.7 MiB/s in  1s ETA:  0s
info: installing component 'rustc'
 59.5 MiB /  59.5 MiB (100 %)  20.3 MiB/s in  2s ETA:  0s
info: installing component 'rustfmt'

  stable-x86_64-pc-windows-msvc installed - (rustc does not exist)

info: checking for self-update
wincowger@Win:/mnt/c/WINDOWS/system32$ rustup toolchain remove stable-x86_64-pc-windows-gnu
info: no toolchain installed for 'stable-x86_64-pc-windows-gnu'
wincowger@Win:/mnt/c/WINDOWS/system32$ rustup install stable-x86_64-pc-windows-gnu
error: DEPRECATED: future versions of rustup will require --force-non-host to install a non-host toolchain as the default.
warning: toolchain 'stable-x86_64-pc-windows-gnu' may not be able to run on this system.
warning: If you meant to build software to target that platform, perhaps try `rustup target add x86_64-pc-windows-gnu` instead?
info: syncing channel updates for 'stable-x86_64-pc-windows-gnu'
info: latest update on 2023-08-03, rust version 1.71.1 (eb26296b5 2023-08-03)
info: downloading component 'cargo'
info: downloading component 'clippy'
info: downloading component 'rust-docs'
 13.6 MiB /  13.6 MiB (100 %)   6.0 MiB/s in  2s ETA:  0s
info: downloading component 'rust-mingw'
info: downloading component 'rust-std'
info: downloading component 'rustc'
 71.4 MiB /  71.4 MiB (100 %)  17.5 MiB/s in  4s ETA:  0s
info: downloading component 'rustfmt'
info: installing component 'cargo'
info: installing component 'clippy'
info: installing component 'rust-docs'
 13.6 MiB /  13.6 MiB (100 %)  13.0 MiB/s in  1s ETA:  0s
info: installing component 'rust-mingw'
info: installing component 'rust-std'
 23.2 MiB /  23.2 MiB (100 %)  20.1 MiB/s in  1s ETA:  0s
info: installing component 'rustc'
 71.4 MiB /  71.4 MiB (100 %)  22.3 MiB/s in  3s ETA:  0s
info: installing component 'rustfmt'

  stable-x86_64-pc-windows-gnu installed - (rustc does not exist)

info: checking for self-update
wincowger@Win:/mnt/c/WINDOWS/system32$ rustup show
Default host: x86_64-unknown-linux-gnu
rustup home:  /home/wincowger/.rustup

installed toolchains
--------------------

stable-x86_64-pc-windows-gnu
stable-x86_64-pc-windows-msvc
stable-x86_64-unknown-linux-gnu (default)

active toolchain
----------------

stable-x86_64-unknown-linux-gnu (default)
rustc 1.71.1 (eb26296b5 2023-08-03)
```
