### 2023-08-09 beniaminogreen/zoomerjoin#70 @beniaminogreen
Hmm. Seems like you are very close. If it's not working already, you could try:
```
rustup default stable-x86_64-pc-windows-gnu
```

To make sure that the x86-windows-gnu toolchain is set as the default toolchain
