### 2023-08-09 #5693 @beniaminogreen
I agree with the comment that "no false positives" might mislead users as to what zoomerjoin can accomplish. I'd still like to flag the fact that zoomerjoin will never identify units that are less similar than the threshold the user selects as matches, but I'll see if there's a way to say this that sets expectations better.

The code for the synthetic dataset and benchmarking can be found in [this vignette](http://beniamino.org/zoomerjoin/articles/benchmarks.html), and I'll include a link in the paper. I also like the suggestion of adding code examples, so I'll add a small example of joining using the Euclidean and Jaccard distances to the article.

I thought that including a contrubuting.md file was enough to satisfy the JOSS requirements, but since both of you have raised this as a concern, I will write up a more detailed set of instructions on contributing or reporting bugs.

Best,
Ben
