### 2023-08-10 b3c7945
added instructions for people looking to file an issue, seek help, contribute
