### 2023-08-11 beniaminogreen/zoomerjoin#70 @wincowgerDEV
Ok, I have the default toolchain set.
```
rustup show
Default host: x86_64-unknown-linux-gnu
rustup home:  /home/wincowger/.rustup

installed toolchains
--------------------

stable-x86_64-pc-windows-gnu (default)
stable-x86_64-pc-windows-msvc
stable-x86_64-unknown-linux-gnu

active toolchain
----------------

stable-x86_64-pc-windows-gnu (default)
(rustc does not exist)
```

However, I am still getting the same error message when I try to install the package. I did realize there may not be as much connection between WSL and Windows as I previously thought so I tried to run RStudio from WSL using this example: https://support.posit.co/hc/en-us/articles/360049776974-Using-RStudio-Server-in-Windows-WSL2, but that is now giving me a bunch of issues too, lots of errors that don't have resolutions when I check online. What is the workflow you usually use to install on a windows machine?
