### 2023-08-11 beniaminogreen/zoomerjoin#70 @beniaminogreen
Shoot, sorry you are still having issues with this!

Just installed Zoomerjoin on my partner's windows computer, which worked after performing the following steps (which are also what I have done in the past).

1) Install rustup-init.exe from [this site](https://forge.rust-lang.org/infra/other-installation-methods.html) and run it
2) run `rustup install stable-x86_64-pc_windows-gnu` to install the correct toolchain
3) Install zoomerjoin with `devtools::install_github("beniaminogreen/zoomerjoin")`

Reading that article it looks as if Rstudio might not be supported for WSL, which could make the install more difficult (although I am not sure this is the case). I just want to confirm that trying to install both Rust and the R package in WSL, or trying to install both in Windows, correct? I don't think that you would be able to install Rust in WSL and the package with Windows and have the two work together, but I am not sure.

I'll try and look into how the install works on WSL, but I am not sure that I will have access to a device with it on anytime soon.

Best,
Ben
