### 2023-08-11 efaba35
added example code for synthetic dataset, example in paper
