+++ paper.bib
@@ -138,4 +138,10 @@
 	author = {Ivan P. Fellegi and Alan B. Sunter},
 	title = {A Theory for Record Linkage},
 	journal = {Journal of the American Statistical Association}
-}
+}@article{dplyr,
+    title = {dplyr: A Grammar of Data Manipulation},
+    author = {Hadley Wickham and Romain François and Lionel Henry and Kirill Müller and Davis Vaughan},
+    year = {2023},
+    note = {R package version 1.1.2},
+    url = {https://CRAN.R-project.org/package=dplyr},
+  }
