### 2023-08-11 #5693 @beniaminogreen
Updating the article to make it responsive to comments above. Specifically: have added code for the synthetic dataset, linked to benchmarking vignette, and added an example use of `euclidean_inner_join`. Also, have removed sentence about no false positives because I agree it could set the wrong expectation for people considering the package.

Added contributing instructions in a separate update to the repo.

Best,
Ben
