### 2023-08-11 beniaminogreen/zoomerjoin#70 @wincowgerDEV
Great! Got it working. Here is what I did. I think my main issue was that I tried to use WSL, it wasn't necessary in this case.

1. go to https://forge.rust-lang.org/infra/other-installation-methods.html, install from this link
![image](https://github.com/beniaminogreen/zoomerjoin/assets/26821843/8feacef5-c557-44c4-a207-f2b2a175721f)
2. in command prompt run rustup install ```stable-x86_64-pc-windows-msvc``` then run ```rustup install stable-x86_64-pc-windows-gnu``` then run ```rustup default stable-x86_64-pc-windows-gnu```.
3. Install zoomerjoin with ```devtools::install_github("beniaminogreen/zoomerjoin")``` in R.
