### 2023-08-13 a1f5529
detailed install instructions for Windows
