+++ README.md
@@ -30,14 +30,27 @@ compiler](https://www.rust-lang.org/tools/install) installed to compile
 this package. After the package is compiled, Rust is no longer required,
 and can be safely uninstalled.
 
-To install Rust on windows, you can use the Rust installation wizard,
-found
-[here](https://forge.rust-lang.org/infra/other-installation-methods.html).
-On Linux or MacOs, you can install Rust with:
+#### Installing Rust on Linux or Mac:
+
+To install Rust on Linux or Mac, you can simply run the following
+snippet in your terminal.
 
 ``` sh
 curl --proto '=https' --tlsv1.2 -sSf https://sh.rustup.rs | sh
 ```
+
+To install Rust on windows, you can use the Rust installation wizard,
+`rustup-init.exe`, found [at this
+site](https://forge.rust-lang.org/infra/other-installation-methods.html).
+Depending on your version of Windows, you may see an error that looks
+something like this:
+
+    error: toolchain 'stable-x86_64-pc-windows-gnu' is not installed
+
+In this case, you should run
+`rustup install stable-x86_64-pc_windows-gnu` to install the missing
+toolchain. If you’re missing another toolchain, simply type this in the
+place of `stable-x86_64-pc_windows-gnu` in the command above.
 
 ### Installing Package from Github:
 
@@ -55,7 +68,7 @@ devtools::install_github("beniaminogreen/zoomerjoin")
 pak::pkg_install("beniaminogreen/zoomerjoin")
 ```
 
-## Loading The Package
+### Loading The Package
 
 Once the package is installed, you can load it into memory as usual by
 typing:
@@ -64,7 +77,7 @@ typing:
 library(zoomerjoin)
 ```
 
-### Usage:
+## Usage:
 
 The flagship feature of zoomerjoins are the jaccard_join and euclidean
 family of functions, which are designed to be near drop-ins for the
@@ -83,7 +96,7 @@ The `jaccard_join` family of functions provide fast fuzzy-joins for
 strings using the Jaccard distance while the `euclidean_join` family
 provides fuzzy-joins for points or vectors using the Euclidean distance.
 
-#### Example: Joining rows of the Database on Ideology, Money in Politics, and Elections
+### Example: Joining rows of the Database on Ideology, Money in Politics, and Elections
 
 (DIME)
 
@@ -169,30 +182,30 @@ join_out <- jaccard_inner_join(corpus_1, corpus_2, n_gram_width=6, n_bands=20, b
 print(Sys.time() - start_time)
 ```
 
-    ## Time difference of 10.00245 secs
+    ## Time difference of 11.47196 secs
 
 ``` r
 print(join_out)
 ```
 
-    ## # A tibble: 188,352 × 4
-    ##         a field.x                                               b field.y       
-    ##     <dbl> <chr>                                             <dbl> <chr>         
-    ##  1    196 committee for responsible government             868477 tmha committe…
-    ##  2  70987 the law offices of lucio g. valdez               918658 law offices o…
-    ##  3  57366 grande communications networks inc. pac account  963865 grande commun…
-    ##  4 257012 esg international inc                           1002173 acg internati…
-    ##  5  68222 ball and weed                                    886213 ball and weed…
-    ##  6  95976 health care center                              1273955 lake health c…
-    ##  7 128542 mid valley democratic club                       899639 tri valley de…
-    ##  8 121580 steelworkers local 760                          1198389 steelworkers …
-    ##  9 268575 mdsa strategic communications                   1196490 strategic com…
-    ## 10 441473 sw construction inc                             1260268 mat construct…
-    ## # ℹ 188,342 more rows
+    ## # A tibble: 183,482 × 4
[…]
