### 2023-08-13 beniaminogreen/zoomerjoin#70 @beniaminogreen
Great, happy you were able to get this working! I just added instructions to disambiguate which installer Windows users should use, and a tip on installing new toolchains to the README in #74.
