### 2023-08-24 beniaminogreen/zoomerjoin#76 @wincowgerDEV
Reviewing
Additional review comments for https://github.com/openjournals/joss-reviews/issues/5693

Usages in the readme should have consistent a naming convention for the functions: https://github.com/beniaminogreen/zoomerjoin#usage

The jaccard_full_join() has parentheses but euclidean_full_join doesn't which makes it appear like the euclidean functions are actually objects.
