### 2023-08-24 beniaminogreen/zoomerjoin#76 @wincowgerDEV
Check that the function examples in help would run if a user has little experience with R.

For example ```jaccard_anti_join()``` requires the user to have ```library(babynames)``` but it really should start with installing that package if the user doesn't have it.
