### 2023-08-24 beniaminogreen/zoomerjoin#76 @wincowgerDEV
https://github.com/beniaminogreen/zoomerjoin#example-joining-rows-of-the-database-on-ideology-money-in-politics-and-elections

It looks like you actually have this dataset loading into zoomerjoin?

Why not read in this instead of calling an object that the user doesn't immediately have access to? Given this link (https://data.stanford.edu/dime) it isn't immediately obvious which is corpus 1 and which is corpus 2.

```zoomerjoin::dime_data```
