### 2023-08-24 beniaminogreen/zoomerjoin#76 @wincowgerDEV
In the vignette I was surprised that ```print(join_out)``` produced a different number of rows than in the example, mine has 47886 while yours has 48106. Is this because of the probabilistic framework?
