### 2023-08-24 beniaminogreen/zoomerjoin#76 @wincowgerDEV
Typo in ```jaacrd_curve``` in the vignette.
