### 2023-08-24 beniaminogreen/zoomerjoin#76 @wincowgerDEV
There seems to be quite a bit of variability in the run time between devices for different functions.

The euclidean_inner_join example is taking much longer on my device than 30 seconds (several minutes already) http://beniamino.org/zoomerjoin/articles/matching_vectors.html

Also join_out example in the vignette took only 3 seconds on my machine.
