### 2023-08-25 beniaminogreen/zoomerjoin#76 @beniaminogreen
Thanks for such detailed feedback on the package. I'll shortly create a branch for processing all these.

It seems that a common thread between all of these  is that the code examples in the function documentations need to be streamlined and checked so they run quickly and without errors.  I'll work on this in the next few days, as well as fixing typos / responding to the other comments.

Based on a comment from the other reviewer, I decided to package a copy of the DIME donor dataset with Zoomerjoin to make the examples more accessible to users. The two corpuses are just random subsets of the donors listed in the database (contrived, but consistent with a previous paper on fuzzy-joining in the field) I'll make this more clear in the documentation as well.

The varying number of observations in the joined dataset is indeed due to the probabilistic algorithm.  For my use cases, I just set the `n_bands` parameter, which governs how many matches can be discarded due to chance, high enough that the results are consistent every run.   This probably isn't ideal from a reproducibility perspective, so I'll create a PR to make the Rust code take the random seed from R so results can be replicated with `set.seed`.

Best, and thanks,
Ben
