### 2023-08-25 ee0f107
shortened euclidean join vignette
