### 2023-08-26 beniaminogreen/zoomerjoin#76 @beniaminogreen
To do for this thread:

- [ ] add babynames install / import to examples or change to use the dataset packaged with zoomerjoin.
 - [ ] make sure all functions are labelled with `function()` not `function` in the documentation
 - [ ] Take the random seed from R and pass to Rust
 - [ ] Set seed in all vingettes so results are reproducible
 - [ ] fix `jaccard_crve` typo
