### 2023-08-26 92cc1d6
deleted ubuntu release workflow because it is not building on GH
