### 2023-08-26 94d8305
added seed to guided tour vignettes
