+++ vignettes/guided_tour.Rmd
@@ -66,6 +66,7 @@ The two Corpuses can't be directly joined because of misspellings. This means
 we must use the fuzzy-matching capabilities of zoomerjoin:
 
 ```{r}
+set.seed(1)
 start_time <- Sys.time()
 join_out <- jaccard_inner_join(corpus_1, corpus_2,
                            by = "field", n_gram_width=6,
