### 2023-08-26 99719b8
added install babynames to jaccard join examples
