+++ man/jaccard_left_join.Rd
@@ -72,7 +72,7 @@ Fuzzy left-join using minihashing
 }
 \examples{
 # load baby names data
-install.packages("babynames")
+#install.packages("babynames")
 library(babynames)
 
 baby_names <- data.frame(name = tolower(unique(babynames$name)))
