+++ man/jaccard_right_join.Rd
@@ -72,7 +72,7 @@ Fuzzy right-join using minihashing
 }
 \examples{
 # load baby names data
-install.packages("babynames")
+#install.packages("babynames")
 library(babynames)
 
 baby_names <- data.frame(name = tolower(unique(babynames$name)))
