### 2023-09-04 350ed7b
added option to turn on progress bar
