### 2023-09-13 #5693 @beniaminogreen
Hi All,

Thanks for your detailed feedback about the package.

In response to some of the comments you've given, I have made the rust code in the package capture R's random seed, so you can set the seed and make sure that all the results are repicable. This helped me catch a locking issue with some of the rust code.

I've bundled a version of the dime dataset with the package so that users can start experimenting with the joining function immediately after installing the package.

It looks like both of you are having difficulty replicating some of the performance claims in the paper. Ideally, I would like to include benchmarks on large datasets to show off the scaling of the package, but I recognize that this makes it tough for replicators. Is there any way I can make it easier for you to run the benchmarks?

Similarly, please don't hesitate to let me know if I can make it easier for you to sign off on any of the other steps.

Best, and many thanks for your help,
Ben
