### 2023-09-13 #5693 @wincowgerDEV
@beniaminogreen, I am trying to run through the examples a last time to see how the update works. When I run your example at the readme.md: https://github.com/beniaminogreen/zoomerjoin

corpus_1 doesn't appear to be part of your package. Could you use dime_data to split corpus 1 and 2 for the user so that they can get up and running without having to guess how they should be split?
