### 2023-09-22 #5693 @beniaminogreen
Thanks for your kind words about the package! I have just completed the checklist below:

**Version Number:** v0.1.0
**Zenodo Release:** 10.5281/zenodo.8370652

[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.8370652.svg)](https://doi.org/10.5281/zenodo.8370652)

Best and thanks for your help,
Ben
